Added new basic routing format

+++ lib/DSP/FilterDelay.cpp
@@ -2,12 +2,12 @@
 #include "Constants.h"
 #include <cmath> //lround
 
-const int FilterDelay::getSampleDelay(const uint32_t sampleRate, double delay, const bool useUnitMeter) {
+const uint32_t FilterDelay::getSampleDelay(const uint32_t sampleRate, double delay, const bool useUnitMeter) {
     //Value is in meter. Convert to milliseconds
     if (useUnitMeter) {
         delay = 1000.0 * delay / SPEED_OF_SOUND;
     }
-    return std::lround(sampleRate * delay / 1000.0);
+    return (uint32_t)std::lround(sampleRate * delay / 1000.0);
 }
 
 FilterDelay::FilterDelay() {
@@ -19,7 +19,7 @@ FilterDelay::FilterDelay(const uint32_t sampleRate, const double delay, const bo
     init(getSampleDelay(sampleRate, delay, useUnitMeter));
 }
 
-FilterDelay::FilterDelay(const int sampleDelay) {
+FilterDelay::FilterDelay(const uint32_t sampleDelay) {
     init(sampleDelay);
 }
 
@@ -27,7 +27,7 @@ FilterDelay::~FilterDelay() {
     delete[] _pBuffer;
 }
 
-void FilterDelay::init(const int sampleDelay) {
+void FilterDelay::init(const uint32_t sampleDelay) {
     _size = sampleDelay;
     _index = 0;
     _pBuffer = new double[_size];
